obs-ffmpeg: Remove hls_deactivate
