+++ plugins/obs-ffmpeg/CMakeLists.txt
@@ -22,8 +22,7 @@ set(obs-ffmpeg_config_HEADERS
 set(obs-ffmpeg_HEADERS
 	obs-ffmpeg-formats.h
 	obs-ffmpeg-compat.h
-	obs-ffmpeg-mux.h
-	obs-ffmpeg-hls-mux.h)
+	obs-ffmpeg-mux.h)
 
 set(obs-ffmpeg_SOURCES
 	obs-ffmpeg.c
