+++ plugins/obs-ffmpeg/ffmpeg-mux/ffmpeg-mux.h
@@ -16,6 +16,7 @@
 
 #pragma once
 
+#include <stdbool.h>
 #include <stdint.h>
 
 enum ffm_packet_type {
